feat: ADD calculator.py 기존 main.py에 있던 계산하는 함수들을 package/calculator.py로 옮기고 main.py에서 calculator.py 불러와 사용
